scripts/md-to-docx/requirements.txt
lxml==5.3.0
python-docx==1.1.2
PyYAML==6.0.2
typing_extensions==4.12.2
python-dotenv
questionary